small inline asm functions updated:  - fcossin in all files now uses proper cosf and sinf functions  - dtol now has better C  - dbound now uses BOUND macro  - scale now has C version
These changes are from older commits, but they never made their way into master

+++ source/game.cpp
@@ -374,8 +374,8 @@ void botinit ()
 static inline void fcossin (float a, float *c, float *s)
 {
 	#if defined(__NOASM__)
-	*c = cos(a);
-	*s = sin(a);
+	*c = cosf(a);
+	*s = sinf(a);
 	#endif
 	#if defined(__GNUC__) && !defined(__NOASM__) //AT&T SYNTAX ASSEMBLY
 	__asm__ __volatile__
